+++ lib/ts/recipe/thirdparty/recipe.ts
@@ -15,7 +15,7 @@
 
 import RecipeModule from "../../recipeModule";
 import { NormalisedAppinfo, APIHandled, RecipeListFunction, HTTPMethod } from "../../types";
-import { TypeInput, TypeNormalisedInput, TypeProvider, RecipeInterface, User, APIInterface } from "./types";
+import { TypeInput, TypeNormalisedInput, TypeProvider, RecipeInterface, APIInterface } from "./types";
 import { validateAndNormaliseUserInput } from "./utils";
 import EmailVerificationRecipe from "../emailverification/recipe";
 import STError from "./error";
@@ -171,17 +171,20 @@ export default class Recipe extends RecipeModule {
         return userInfo.email;
     };
 
-    createEmailVerificationToken = async (userId: string): Promise<string> => {
-        return this.emailVerificationRecipe.createEmailVerificationToken(userId, await this.getEmailForUserId(userId));
+    createEmailVerificationToken = async (userId: string) => {
+        return this.emailVerificationRecipe.recipeInterfaceImpl.createEmailVerificationToken({
+            userId,
+            email: await this.getEmailForUserId(userId),
+        });
     };
 
-    verifyEmailUsingToken = async (token: string): Promise<User> => {
-        let user = await this.emailVerificationRecipe.verifyEmailUsingToken(token);
-        let userInThisRecipe = await this.recipeInterfaceImpl.getUserById({ userId: user.id });
-        if (userInThisRecipe === undefined) {
-            throw Error("Unknown User ID provided");
+    verifyEmailUsingToken = async (token: string) => {
+        let response = await this.emailVerificationRecipe.recipeInterfaceImpl.verifyEmailUsingToken({ token });
+        if (response.status === "OK") {
+            let userInThisRecipe = await this.recipeInterfaceImpl.getUserById({ userId: response.user.id });
+            return userInThisRecipe;
         }
-        return userInThisRecipe;
+        return response;
     };
 
     isEmailVerified = async (userId: string) => {